Rename patientCard module to medicalHistory; move PatientHistory contents into module root and rename component to MedicalHistory
(cherry picked from commit d0a15d1bf40b0147cdc052d35d7eaf0fd36143ba)
