+++ frontend/src/modules/medicalHistory/MedicalHistory.script.js
@@ -1,19 +1,19 @@
 import { extractIdFromIri } from '@/modules/shared/utils/apiHelpers';
-import { useAppointmentAddStore } from '@/modules/patientCard/stores/appointmentAddStore';
-import AppointmentAdd from '@/modules/patientCard/components/PatientHistory/AppointmentAdd/AppointmentAdd.vue';
-import TestAddModal from '@/modules/patientCard/components/PatientHistory/TestAddModal/TestAddModal.vue';
-import PatientCard from '@/modules/patientCard/components/PatientHistory/PatientCard/PatientCard.vue';
-import TreatmentCard from '@/modules/patientCard/components/PatientHistory/TreatmentCard/TreatmentCard.vue';
-import { usePatientCardStore } from '@/modules/patientCard/stores/patientCardStore';
-import { useTreatmentStore } from '@/modules/patientCard/stores/treatmentStore';
-import { useMedicalTableStore } from '@/modules/patientCard/stores/medicalTableStore';
-import MedicalTable from '@/modules/patientCard/components/PatientHistory/MedicalTable/MedicalTable.vue';
-import Pharmacogenetics from '@/modules/patientCard/components/PatientHistory/Pharmacogenetics/Pharmacogenetics.vue';
-import VitalsCard from '@/modules/patientCard/components/VitalsCard/VitalsCard.vue';
-import RiskScale from '@/modules/patientCard/components/RiskScale/RiskScale.vue';
+import { useAppointmentAddStore } from '@/modules/medicalHistory/stores/appointmentAddStore';
+import AppointmentAdd from '@/modules/medicalHistory/AppointmentAdd/AppointmentAdd.vue';
+import TestAddModal from '@/modules/medicalHistory/TestAddModal/TestAddModal.vue';
+import PatientCard from '@/modules/medicalHistory/PatientCard/PatientCard.vue';
+import TreatmentCard from '@/modules/medicalHistory/TreatmentCard/TreatmentCard.vue';
+import { usePatientCardStore } from '@/modules/medicalHistory/stores/patientCardStore';
+import { useTreatmentStore } from '@/modules/medicalHistory/stores/treatmentStore';
+import { useMedicalTableStore } from '@/modules/medicalHistory/stores/medicalTableStore';
+import MedicalTable from '@/modules/medicalHistory/MedicalTable/MedicalTable.vue';
+import Pharmacogenetics from '@/modules/medicalHistory/Pharmacogenetics/Pharmacogenetics.vue';
+import VitalsCard from '@/modules/medicalHistory/components/VitalsCard/VitalsCard.vue';
+import RiskScale from '@/modules/medicalHistory/components/RiskScale/RiskScale.vue';
 
 export default {
-    name: 'PatientHistory',
+    name: 'MedicalHistory',
     components: { 
         RiskScale, 
         AppointmentAdd,
