+++ frontend/src/modules/medicalHistory/MedicalHistory.vue
@@ -1,0 +1,3 @@
+<template src="./MedicalHistory.template.html"></template>
+<script src="./MedicalHistory.script.js"></script>
+<style src="./MedicalHistory.style.css"></style>
