+++ frontend/src/modules/medicalHistory/MedicalTable/MedicalTable.style.css
