+++ frontend/src/modules/medicalHistory/MedicalTable/MedicalTable.template.html
