+++ frontend/src/modules/medicalHistory/MedicalTable/MedicalTable.vue
