+++ frontend/src/modules/medicalHistory/MnoChart/MnoChart.script.js
