+++ frontend/src/modules/medicalHistory/MnoChart/MnoChart.style.css
