+++ frontend/src/modules/medicalHistory/MnoChart/MnoChart.vue
