+++ frontend/src/modules/medicalHistory/PatientCard/PatientCard.style.css
