+++ frontend/src/modules/medicalHistory/PatientCard/PatientCard.template.html
