+++ frontend/src/modules/medicalHistory/PatientCard/PatientCard.vue
