+++ frontend/src/modules/medicalHistory/Pharmacogenetics/Pharmacogenetics.style.css
