+++ frontend/src/modules/medicalHistory/Pharmacogenetics/Pharmacogenetics.vue
