+++ frontend/src/modules/medicalHistory/TreatmentCard/TreatmentCard.style.css
