+++ frontend/src/modules/medicalHistory/TreatmentCard/TreatmentCard.template.html
