+++ frontend/src/modules/medicalHistory/TreatmentCard/TreatmentCard.vue
