+++ frontend/src/modules/medicalHistory/components/RiskScale/RiskScale.script.js
