+++ frontend/src/modules/medicalHistory/components/RiskScale/RiskScale.style.css
