+++ frontend/src/modules/medicalHistory/components/RiskScale/RiskScale.template.html
