+++ frontend/src/modules/medicalHistory/components/RiskScale/RiskScale.vue
