+++ frontend/src/modules/medicalHistory/components/VitalsCard/VitalsCard.style.css
