+++ frontend/src/modules/medicalHistory/components/VitalsCard/VitalsCard.template.html
