+++ frontend/src/modules/medicalHistory/components/VitalsCard/VitalsCard.vue
