+++ frontend/src/modules/medicalHistory/stores/medicalTableStore.js
