+++ frontend/src/modules/medicalHistory/stores/patientCardStore.js
