+++ frontend/src/modules/medicalHistory/stores/patientVitalsLatestStore.js
