--- frontend/src/modules/patientCard/components/PatientHistory/PatientHistory.vue
@@ -1,3 +1,0 @@
-<template src="./PatientHistory.template.html"></template>
-<script src="./PatientHistory.script.js"></script>
-<style src="./PatientHistory.style.css"></style>
